Extract Launcher, implement current_url
