+++ lib/shimmer.rb
@@ -2,6 +2,7 @@ require "shimmer/version"
 require "shimmer/driver"
 require "shimmer/browser"
 require "shimmer/node"
+require "shimmer/launcher"
 
 module Shimmer
 end
